Drop unnecessary 'init;'s
The `init;` accessor declaration is only required if the property
is assigned to during initialization of an object *outside* the
constructor for the type the property is declared on.

e.g. during `new Foo { X = 42 }`, `foo with { X = 42 }` or in a
derived class's constructor.

(See the C# language reference for technical details:
https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/proposals/csharp-9.0/init#init-accessors
)

Since these usages don't apply to any of the properties changed
in this commit, it is safe to drop them.